+++ src/main/java/bungee/me/RockinChaos/cloudsync/listeners/Chat.java
@@ -1,4 +1,4 @@
-package src.main.java.me.RockinChaos.cloudsync.bungee.listeners;
+package bungee.me.RockinChaos.cloudsync.listeners;
 
 import net.md_5.bungee.api.event.ChatEvent;
 import net.md_5.bungee.api.plugin.Listener;
