+++ src/main/resources/velocity-plugin.json
@@ -5,5 +5,5 @@
   "url": "https://www.spigotmc.org/resources/cloudsync.93382/",
   "authors": [ "RockinChaos" ],
   "dependencies": [],
-  "main": "src.main.java.me.RockinChaos.cloudsync.velocity.CloudSync"
+  "main": "velocity.me.RockinChaos.cloudsync.CloudSync"
 }
